+++ src/theme/index.js
@@ -1,2 +1,2 @@
-export { default } from './theme'
-export * from './theme'
+export { default as createTheme } from './theme'
+export { default as invertTheme } from './invertTheme'