+++ src/theme/invertTheme.js
@@ -1,0 +1,18 @@
+import { merge } from 'lodash-es'
+
+const invertTheme = (theme = {}) => {
+  const { palette = {} } = theme
+  const { primary = {}, secondary = {} } = palette
+  return merge(theme, {
+    palette: {
+      primary: {
+        main: secondary.main,
+      },
+      secondary: {
+        main: primary.main,
+      },
+    },
+  })
+}
+
+export default invertTheme