+++ src/components/Card/Card.js
@@ -1,7 +1,7 @@
 import React, { forwardRef } from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
-import { isEmpty, isString } from 'lodash-es'
+import { isEmpty, isString } from 'utils'
 
 import { makeStyles } from '@material-ui/core/styles'
 import {