+++ src/components/CookiesConsent/components/Category.js
@@ -9,7 +9,7 @@ import {
   Typography,
 } from '@material-ui/core'
 import { makeStyles } from '@material-ui/core/styles'
-import { isBoolean } from 'lodash-es'
+import { isBoolean } from 'utils'
 
 import { useIsSmallScreen } from 'hooks'