+++ src/components/Footer/Footer.js
@@ -1,7 +1,7 @@
 import React, { forwardRef } from 'react'
 import styled from 'styled-components'
 import PropTypes from 'prop-types'
-import { isEmpty } from 'lodash-es'
+import { isEmpty } from 'utils'
 
 import BottomBanner from './components/BottomBanner'
 import Column from './components/Column'