+++ src/components/Navbar/components/Brand.js
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import PropTypes from 'prop-types'
 import { Link, Typography } from '@material-ui/core'
 import { makeStyles } from '@material-ui/core/styles'
-import { isEmpty } from 'lodash-es'
+import { isEmpty } from 'utils'
 
 const useStyles = makeStyles({
   title: {