+++ src/utils/isEmpty.js
@@ -1,0 +1,1 @@
+export default (obj = {}) => obj.length < 1 || Object.keys(obj).length < 1