+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 * Fix label not changing value of switch in `CookiesConsent`. ([@alexandre-lelain](https://github.com/alexandre-lelain))
 
+## Chores
+
+* Bump dependencies to latest versions. ([@alexandre-lelain](https://github.com/alexandre-lelain))
+
 # 2.3.4
 
 ## Docs